Format Case and Vaccination Parameters for Model Input (Including Pre-1980)

Wrapper function that prepares all time-varying parameters for use in an age-, vaccine-, and risk-structured infectious disease model, including extrapolated vaccination coverage prior to 1980 if available.

Usage

case_vaccine_to_param(
  demog_data,
  processed_vaccination,
  processed_vaccination_sia,
  processed_case,
  vaccination_schedule,
  vaccination_pre1980,
  vaccine_parameters,
  disease_parameters,
  WHO_seed_switch,
  custom_routine_vaccination = NA,
  custom_sia_vaccination = NA
)

Arguments

demog_data

Output list from `process_demography()`.

processed_vaccination

Data frame of routine vaccine coverage.

processed_vaccination_sia

Data frame of SIA coverage.

processed_case

WHO case reports processed for the country of interest.

vaccination_schedule

Full WHO vaccine schedule data frame.

vaccination_pre1980

Data frame of historical vaccination introduction assumptions.

vaccine_parameters

Data frame of vaccination parameters.

custom_routine_vaccination

Data frame or NA indicating custom routine vaccination.

custom_sia_vaccination

Data frame or NA indicating custom sia vaccination.

Value

A named list of parameters including coverage arrays and seeding inputs.